Add copyright and update Config files to match
