+++ exporters/metric/cortex/go.mod
@@ -3,14 +3,13 @@ module go.opentelemetry.io/contrib/exporters/metric/cortex
 go 1.14
 
 require (
+	github.com/gogo/protobuf v1.3.1 // indirect
 	github.com/golang/protobuf v1.4.2
 	github.com/golang/snappy v0.0.1
 	github.com/google/go-cmp v0.5.1
+	github.com/grpc-ecosystem/grpc-gateway v1.14.6 // indirect
 	github.com/prometheus/prometheus v2.5.0+incompatible
-	github.com/spf13/afero v1.3.3
-	github.com/spf13/viper v1.7.1
 	github.com/stretchr/testify v1.6.1
 	go.opentelemetry.io/otel v0.10.0
 	go.opentelemetry.io/otel/sdk v0.10.0
-	google.golang.org/protobuf v1.23.0
 )
